+++ config/dev.exs
@@ -7,7 +7,6 @@ use Mix.Config
 # watchers to your application. For example, we use it
 # with brunch.io to recompile .js and .css sources.
 config :td_dd, TdDdWeb.Endpoint,
-  http: [port: 4005],
   debug_errors: true,
   code_reloader: true,
   check_origin: false,